Move envs in tdm to base
