+++ garage/torch/envs/tdm/half_cheetah_env.py
@@ -7,7 +7,7 @@ from garage.torch.core.eval_util import create_stats_ordered_dict
 from garage.torch.core.serializable import Serializable
 from garage.torch.core import logger as default_logger
 from garage.torch.samplers.util import get_stat_in_paths
-from garage.torch.algos.tdm.envs.multitask_env import MultitaskEnv
+from garage.torch.envs.tdm.multitask_env import MultitaskEnv
 
 MAX_SPEED = 6
